+++ Assets/Inworld.AI/Scripts/Runtime/Audio/AudioCapture.cs
@@ -22,25 +22,26 @@ namespace Inworld
         public UnityEvent OnRecordingStart;
         public UnityEvent OnRecordingEnd;
         public bool IsCapturing { get; set; }
-        [SerializeField] int m_AudioRate = 16000;
-        [SerializeField] int m_BufferSeconds = 1;
+        [SerializeField] protected int m_AudioRate = 16000;
+        [SerializeField] protected int m_BufferSeconds = 1;
         
         readonly List<ByteString> m_AudioToPush = new List<ByteString>();
         // Size of audioclip used to collect information, need to be big enough to keep up with collect. 
         int m_BufferSize;
         const int k_SizeofInt16 = sizeof(short);
         byte[] m_ByteBuffer;
-        float[] m_FloatBuffer;
-        AudioClip m_Recording;
+        protected float[] m_FloatBuffer;
+        protected AudioClip m_Recording;
         float m_CDCounter;
         // Last known position in AudioClip buffer.
         int m_LastPosition;
         bool m_AutoPush;
+        protected string m_CurrentDevice = null;
 
         public void StartRecording(bool autoPush = true)
         {
-            if (!Microphone.IsRecording(null))
-                m_Recording = Microphone.Start(null, true, m_BufferSeconds, m_AudioRate);
+            if (!Microphone.IsRecording(m_CurrentDevice))
+                m_Recording = Microphone.Start(m_CurrentDevice, true, m_BufferSeconds, m_AudioRate);
             m_LastPosition = Microphone.GetPosition(null);
             m_AudioToPush.Clear();
             IsCapturing = true;
@@ -55,26 +56,26 @@ namespace Inworld
             m_AutoPush = true;
             OnRecordingEnd.Invoke();
         }
-        void Awake()
+        protected virtual void Awake()
         {
             m_BufferSize = m_BufferSeconds * m_AudioRate;
             m_ByteBuffer = new byte[m_BufferSize * 1 * k_SizeofInt16];
             m_FloatBuffer = new float[m_BufferSize * 1];
         }
-        void Start()
+        protected virtual void Start()
         {
-            m_Recording = Microphone.Start(null, true, m_BufferSeconds, m_AudioRate);
+            m_Recording = Microphone.Start(m_CurrentDevice, true, m_BufferSeconds, m_AudioRate);
         }
         void Update()
         {
             if (!IsCapturing)
                 return;
-            if (!Microphone.IsRecording(null))
+            if (!Microphone.IsRecording(m_CurrentDevice))
                 StartRecording();
             if (m_CDCounter <= 0)
             {
                 m_CDCounter = 0.1f;
-                _Collect();
+                Collect();
             }
             m_CDCounter -= Time.deltaTime;
         }
@@ -82,17 +83,24 @@ namespace Inworld
         {
             StopRecording();
         }
-        void _Collect()
+        protected int GetAudioData()
         {
-            int nPosition = Microphone.GetPosition(null);
+            int nPosition = Microphone.GetPosition(m_CurrentDevice);
             if (nPosition < m_LastPosition)
                 nPosition = m_BufferSize;
             if (nPosition <= m_LastPosition)
-                return;
+                return -1;
             int nSize = nPosition - m_LastPosition;
             if (!m_Recording.GetData(m_FloatBuffer, m_LastPosition))
-                return;
+                return -1;
             m_LastPosition = nPosition % m_BufferSize;
+            return nSize;
+        }
+        protected virtual void Collect()
+        {
+            int nSize = GetAudioData(); // YAN: Out to m_FloatBuffer.
+            if (nSize < 0)
+                return;
             WavUtility.ConvertAudioClipDataToInt16ByteArray(m_FloatBuffer, nSize * m_Recording.channels, m_ByteBuffer);
             ByteString audioData = ByteString.CopyFrom(m_ByteBuffer, 0, nSize * m_Recording.channels * k_SizeofInt16);
             if (m_AutoPush)