+++ Assets/Inworld.AI/Scripts/Runtime/Data/Settings/InworldLog.cs
@@ -6,34 +6,9 @@
 *************************************************************************************************/
 using System;
 using System.Diagnostics;
-using System.Text.RegularExpressions;
 using Debug = UnityEngine.Debug;
 namespace Inworld.Util
 {
-    public class InworldError
-    {
-        public string statusCode;
-        public string detail;
-        const string k_Pattern = @"StatusCode=""([^""]*)"", Detail=""([^""]*)""";
-        public static InworldError FromString(string error)
-        {
-            InworldError result = new InworldError();
-            Match match = Regex.Match(error, k_Pattern);
-
-            if (match.Success && match.Groups.Count > 1)
-            {
-                result.statusCode = match.Groups[1].Value;
-                result.detail = match.Groups[2].Value;
-            }
-            else
-            {
-                result.statusCode = "Error";
-                result.detail = error;
-            }
-            return result;
-        }
-        public string Message => $"{statusCode}: {detail}";
-    }
     /// <summary>
     ///     Inworld Use UnityEngine's original Log system,
     ///     and does not send any data to our server.